% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_params.R
\name{validate_params}
\alias{validate_params}
\title{Validate Parameters}
\usage{
validate_params(
  parameters,
  types,
  input_files = c(),
  output_files = c(),
  overwrite = c()
)
}
\arguments{
\item{parameters}{list of parameters to knit Rmd with.}

\item{types}{named list, with the same names as `parameters`, noting what
each parameters type should be: "string", "logical", or "double".}

\item{input_files}{list of parameter names that correspond to input files.
Will be checked for existence.}

\item{output_files}{list of parameter names that correspond to output files.
Unless on `overwrite` list, will stop user from overwriting these files.}

\item{overwrite}{list of parameter names that correspond to output files
that can be overwriten.}
}
\description{
Validate Parameters
}
