% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_Rmd.R
\name{write_Rmd}
\alias{write_Rmd}
\title{Writes the Rmd file in the specified directory}
\usage{
write_Rmd(params, output_file = "report", output_dir = ".")
}
\arguments{
\item{params}{A list of named parameters that override custom params specified within the YAML front-matter.}

\item{output_file}{The name of the output file.}

\item{output_dir}{The output directory for the rendered `output_file`.}
}
\value{
path of Rmd file
}
\description{
Writes the Rmd file in the specified directory
}
